added prints

+++ train.py
@@ -129,7 +129,8 @@ def train(input_path, n_epochs):
                     # print(lls_offset_pred.dtype, lls_offset.dtype)
                     print(lls_offset)
                     print("gt", lls_offset.type(torch.LongTensor))
-                    loss = loss_fn(lls_offset_pred, lls_offset.type(torch.LongTensor))
+                    #loss = loss_fn(lls_offset_pred, lls_offset.type(torch.LongTensor))
+                    loss = loss_fn(lls_offset_pred, lls_offset)
                     print("Loss: ", loss)
                     train_total_loss += loss
                     optimizer.zero_grad()
